parseq_jawi/requirements/core.txt
--extra-index-url https://download.pytorch.org/whl/cpu

aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
jinja2==3.1.3
joblib==1.3.2
lightning-utilities==0.10.1
markupsafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pytorch-lightning==2.2.0.post0
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sympy==1.12
timm==0.9.16
torch==2.2.1+cpu
torchmetrics==1.3.1
torchvision==0.17.1+cpu
tqdm==4.66.2
typing-extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
protobuf==5.29.4

# The following packages are considered to be unsafe in a requirements file:
setuptools==69.1.0